web/requirements.txt
Flask
flask-jwt-extended
mysql-connector-python
python-dotenv
yagmail[all]
